requirements.txt
streamlit
openai
python-dotenv
azure-cosmos==4.6.0  
